forge/pyproject.toml
[tool.poetry]
name = "AutoGPT-Forge"
version = "0.1.0"
description = ""
authors = ["Craig Swift <craigswift13@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "forge" }]

[tool.poetry.dependencies]
python = "^3.10"
agbenchmark = { path = "../benchmark", optional = true }
# agbenchmark = {git = "https://github.com/Significant-Gravitas/AutoGPT.git", subdirectory = "benchmark", optional = true}
aiohttp = "^3.8.5"
anthropic = "^0.25.1"
beautifulsoup4 = "^4.12.2"
boto3 = "^1.33.6"
bs4 = "^0.0.1"
charset-normalizer = "^3.1.0"
chromadb = "^0.4.10"
click = "*"
colorama = "^0.4.6"
demjson3 = "^3.0.0"
docker = "*"
duckduckgo-search = "^5.0.0"
fastapi = "^0.109.1"
gitpython = "^3.1.32"
google-api-python-client = "*"
google-cloud-storage = "^2.13.0"
jinja2 = "^3.1.2"
jsonschema = "*"
litellm = "^1.17.9"
openai = "^1.7.2"
Pillow = "*"
playsound = "~1.2.2"
pydantic = "*"
python-docx = "*"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.7"
pylatexenc = "*"
pypdf = "^3.1.0"
pyyaml = "^6.0"
requests = "*"
selenium = "^4.13.0"
sqlalchemy = "^2.0.19"
sentry-sdk = "^1.40.4"
spacy = "^3.0.0"
tenacity = "^8.2.2"
tiktoken = "^0.5.0"
toml = "^0.10.2"
uvicorn = "^0.23.2"
webdriver-manager = "^4.0.1"

[tool.poetry.extras]
benchmark = ["agbenchmark"]

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.3.0"
pre-commit = "^3.3.3"
mypy = "^1.4.1"
flake8 = "^6.0.0"
types-requests = "^2.31.0.2"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
watchdog = "^3.0.0"
mock = "^5.1.0"
autoflake = "^2.2.0"
pydevd-pycharm = "^233.6745.319"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
packages = ["autogpt"]
extend-exclude = '(/dist|/.venv|/venv|/build|/agent|agbenchmark/challenges)/'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip_glob = [".tox", "__pycache__", "*.pyc", "venv*/*", "reports", "venv", "env", "node_modules", ".env", ".venv", "dist", "agent/*", "agbenchmark/challenges/*"]
